feat: region-driven analysis -- statistics, histogram, profiles, plots (M6-22 ... M6-26)
DS9 puts an Analysis menu on every region's dialog, and all five entries
start from the same question: which pixels are inside this region.
`regions/region_analysis.py` answers it once and the rest follows.

The statistics are DS9's, field for field, from `Base::markerAnalysisStats`
(`tksao/frame/frblt.C`): sum, error as the root of its absolute value, area,
surface brightness and its error; then sum, npix, mean, median, min, max,
variance, stddev and rms. The variance is DS9's population form,
`|sum2/n - (sum/n)^2|`, not the sample one, so the numbers match what DS9
prints for the same region -- there is a test comparing it against
`np.var`, which is the same convention.

A shape with annuli reports one row per ring, numbered from one, as DS9
does. Reporting all its rings together would say nothing a plain circle
does not, and it is what makes the radial profile a profile.

Which entries a shape gets is DS9's rule, not a uniform list: a radial
profile needs annuli, a 2D cut needs a line, and a point or a text region
gets no Analysis menu at all. An entry that could only ever answer "this
shape cannot do that" is not offered.

Two things worth recording:

  - The mask is half-open on a box's upper edge. A ten-pixel box centred on
    a pixel centre reaches exactly the centres five away on both sides, and
    counting both gives eleven columns for a ten-pixel box -- a 28%
    overstatement of the area every surface brightness is divided by.
    `Box.contains` stays inclusive, because that is hit-testing, not area.
  - Circles, ellipses and boxes are masked arithmetically rather than
    through `contains`. A per-pixel Python call over a few hundred pixels
    square is a hundred thousand of them, and the Auto toggles remeasure on
    every drag.

`analysis/statistics.py` and `analysis/histogram.py` are adopted rather than
duplicated -- both already took a mask, which is exactly what a region is.
Their two orphan entries are gone, leaving `test_no_orphan_modules` empty of
M6 debts.

The Auto Plot toggles (M6-26) now do what they are for: with one on, a region
drawn or changed opens and keeps up to date that window, so a projection
dragged across a source shows the cut changing rather than needing a dialog
reopened each time.

Fixed on the way, all surfaced by mypy once these modules were reachable: an
integer contour coordinate returned where floats were promised, a rank-any
shape assigned to a two-tuple, and a numpy index used unconverted.

+++ ncrads9/analysis/contour.py
@@ -196,7 +196,10 @@ class ContourGenerator:
             dilated = ndimage.binary_dilation(binary)
             edge = dilated ^ binary
             y_coords, x_coords = np.where(edge)
-            contours.append([(x_coords, y_coords)])
+            # `np.where` gives integer indices; the caller is promised
+            # floating-point coordinates, since a real contour lies between
+            # pixels.
+            contours.append([(x_coords.astype(float), y_coords.astype(float))])
 
         return contours
